[REFACTOR] : ♻️ 업데이트 된 공용 컴포넌트를 사용하도록 수정
